feat: TNL-9460 fix failing test
contentstore/views test
